refactor: Add support for additional asset image source
This commit adds support for fetching asset images from the 'icons.llamao.fi' source in addition to the existing 'xdefi-static.s3.eu-west-1.amazonaws.com' source. This allows for a wider range of asset images to be displayed in the application.

The 'Asset' component has been updated to conditionally render the new 'AssetImageEvm' component when the asset has a 'chainName' property. The 'AssetImageEvm' component displays the asset image along with the logo of the corresponding EVM chain.

The 'getChainLogoByName' function has been added to the 'chainLogos.ts' file to retrieve the URL of the EVM chain logo based on the chain name.

These changes improve the visual representation of assets in the application and enhance the user experience.